# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kuben API",
    "description": "Kuben — Kubernetes-native PaaS control plane.",
    "license": {
      "name": "Apache-2.0"
    },
    "version": "1.1.2"
  },
  "paths": {
...
    "/api/v1/projects/{project}/environments/{environment}/apps": {
      "get": {
        "tags": [
          "apps"
        ],
        "summary": "List the apps of an environment.",
        "operationId": "listApps",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "description": "Project name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "environment",
            "in": "path",
            "description": "Environment short name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "apps"
        ],
        "summary": "Deploy a new app from a container image, or from a Git repository. A tag\nis resolved to a digest at its registry; a Git app deploys with its first\nbuild.",
        "operationId": "createApp",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "description": "Project name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "environment",
            "in": "path",
            "description": "Environment short name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApp"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "The image's registry cannot be reached",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project}/environments/{environment}/apps/{app}": {
      "get": {
        "tags": [
          "apps"
        ],
        "summary": "App detail with pods. Plain env values are included only for callers\nholding `secret-read`.",
        "operationId": "getApp",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "description": "Project name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "environment",
            "in": "path",
            "description": "Environment short name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "app",
            "in": "path",
            "description": "App name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppDetail"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "apps"
        ],
        "summary": "Delete an app and everything it owns. Volumes are kept unless\n`delete_volumes=true`.",
        "operationId": "deleteApp",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "description": "Project name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "environment",
            "in": "path",
            "description": "Environment short name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "app",
            "in": "path",
            "description": "App name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "delete_volumes",
            "in": "query",
            "description": "Also delete the app's persistent volumes. Irreversible.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "apps"
        ],
        "summary": "Update an app (image changes require `app-deploy`). Every change is a new\ndeployment run; a new tag is resolved to a digest at its registry.",
        "operationId": "updateApp",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "description": "Project name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "environment",
            "in": "path",
            "description": "Environment short name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "app",
            "in": "path",
            "description": "App name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateApp"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "The image's registry cannot be reached",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{project}/environments/{environment}/apps/{app}/deployments": {
      "get": {
        "tags": [
          "apps"
        ],
        "summary": "The app's newest deployment runs, each with its timeline.",
        "operationId": "listDeployments",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "description": "Project name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "environment",
            "in": "path",
            "description": "Environment short name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "app",
            "in": "path",
            "description": "App name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "How many runs, newest first (1–50, default 10).",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeploymentSummary"
                  }
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "apps"
        ],
        "summary": "Accept a deployment of this app.",
        "description": "The release (from `image`, or an existing `release`), the configuration\nrevision and the run are written in one transaction with the audit record\nand the message to the executors; then the call answers `202` with the\nrun's `Location`. Replaying the request with the same `Idempotency-Key`\nreturns the first run.",
        "operationId": "startDeployment",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "description": "Project name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "environment",
            "in": "path",
            "description": "Environment short name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "app",
            "in": "path",
            "description": "App name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Replays of the same request return the first run",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartDeploymentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted: poll the `Location`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "A stale expected generation, or an Idempotency-Key used for another request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{project}/environments/{environment}/apps/{app}/handover": {
      "post": {
        "tags": [
          "apps"
        ],
        "summary": "Hand an app over from the App controller to its cluster's agent: its\nnext run goes through the agent, which adopts the app's workloads in\nplace (no new pods), and every later run follows. Only toward the agent;\nthe cluster needs a linked agent that carries applications.",
        "operationId": "handOverApp",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "description": "Project name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "environment",
            "in": "path",
            "description": "Environment short name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "app",
            "in": "path",
            "description": "App name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Handover scheduled"
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Delivered by the agent already, being deleted, or no agent to take it",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AppDto": {
        "type": "object",
        "required": [
          "name",
          "project",
          "environment",
          "namespace",
          "ready",
          "processes",
          "env",
          "domains",
          "volumes"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "project": {
            "type": "string"
          },
          "environment": {
            "type": "string",
            "description": "Environment short name."
          },
          "namespace": {
            "type": "string"
          },
          "image": {
            "type": [
              "string",
              "null"
            ]
          },
          "git_repo": {
            "type": [
              "string",
              "null"
            ]
          },
          "url": {
            "type": [
              "string",
              "null"
            ]
          },
          "ready": {
            "type": "boolean"
          },
          "reason": {
            "type": [
              "string",
              "null"
            ],
            "description": "`Available`, `Progressing`, `RolloutFailed`, `Scheduled`, `AwaitingBuild`, ..."
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          },
          "processes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProcessDto"
            }
          },
          "env": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EnvVarDto"
            },
            "description": "Values are only included in the app detail, for callers allowed to read secrets."
          },
          "domains": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "volumes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VolumeDto"
            }
          },
          "created_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "exposure": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ExposureDto",
                "description": "Whether the app can be reached through the gateway, apart from\nwhether it runs (null while it has no route)."
              }
            ]
          },
          "paused": {
            "type": [
              "string",
              "null"
            ],
            "description": "Why delivery is paused, while it is: new runs wait."
          }
        }
      },
...
    }
  },
...
}